[Cpp API Compatibility] move implementation from Utils.h to Utils.cpp
