+++ paddle/phi/api/include/compat/ATen/Utils.h
@@ -14,82 +14,23 @@
 
 #pragma once
 
-#include <ATen/ops/empty.h>
-#include <c10/core/Layout.h>
-#include <c10/core/ScalarType.h>
-#include <c10/util/ArrayRef.h>
-#include <c10/util/Exception.h>
-#include <c10/util/accumulate.h>
-
-#include <algorithm>
-
-#include "paddle/phi/api/include/sparse_api.h"
-#include "paddle/phi/api/include/tensor.h"
+#include <ATen/core/Tensor.h>
 
 namespace at {
 namespace detail {
 
 template <typename T>
-Tensor tensor_cpu(ArrayRef<T> values, const TensorOptions& options) {
-  constexpr auto native_scalar_type = c10::CppTypeToScalarType<T>::value;
-  auto result = at::empty(values.size(), options.dtype(native_scalar_type));
-  PD_CHECK(result.is_contiguous());
-  std::copy(values.begin(), values.end(), result.template data_ptr<T>());
-  if (options.dtype() != native_scalar_type) {
-    return result.to(at::TensorOptions().dtype(options.dtype()));
-  }
-  return result;
-}
+Tensor tensor_cpu(ArrayRef<T> values, const TensorOptions& options);
 
 template <typename T>
-Tensor tensor_backend(ArrayRef<T> values, const TensorOptions& options) {
-  auto cpu_tensor =
-      tensor_cpu(values, options.device(c10::Device(c10::DeviceType::CPU)));
-  return cpu_tensor.to(options.device());
-}
+Tensor tensor_backend(ArrayRef<T> values, const TensorOptions& options);
 
 template <typename T>
-Tensor tensor_complex_cpu(ArrayRef<T> values, const TensorOptions& options) {
-  constexpr auto native_scalar_type = c10::CppTypeToScalarType<T>::value;
-  auto result = at::empty(values.size(), options.dtype(native_scalar_type));
-  PD_CHECK(result.is_contiguous());
-  std::copy(values.begin(), values.end(), result.template data_ptr<T>());
-  if (options.dtype() != native_scalar_type) {
-    return result.to(at::TensorOptions().dtype(options.dtype()));
-  }
-  return result;
-}
+Tensor tensor_complex_cpu(ArrayRef<T> values, const TensorOptions& options);
 
 template <typename T>
-Tensor tensor_complex_backend(ArrayRef<T> values,
-                              const TensorOptions& options) {
-  auto cpu_tensor = tensor_complex_cpu(
-      values, options.device(c10::Device(c10::DeviceType::CPU)));
-  return cpu_tensor.to(options.device());
-}
+Tensor tensor_complex_backend(ArrayRef<T> values, const TensorOptions& options);
 
 }  // namespace detail
-
-#define TENSOR(T, _1)                                                      \
-  inline Tensor tensor(ArrayRef<T> values, const TensorOptions& options) { \
-    if (options.device().type() != c10::DeviceType::CPU) {                 \
-      return at::detail::tensor_backend(values, options);                  \
-    } else {                                                               \
-      return at::detail::tensor_cpu(values, options);                      \
-    }                                                                      \
-  }
-AT_FORALL_SCALAR_TYPES_AND3(Bool, Half, BFloat16, TENSOR)
-#undef TENSOR
-
-#define TENSOR(T, _1)                                                      \
-  inline Tensor tensor(ArrayRef<T> values, const TensorOptions& options) { \
-    if (options.device().type() != c10::DeviceType::CPU) {                 \
-      return at::detail::tensor_complex_backend(values, options);          \
-    } else {                                                               \
-      return at::detail::tensor_complex_cpu(values, options);              \
-    }                                                                      \
-  }
-AT_FORALL_COMPLEX_TYPES(TENSOR)
-#undef TENSOR
 
 }  // namespace at
