+++ paddle/phi/api/include/compat/CMakeLists.txt
@@ -2,5 +2,6 @@ collect_srcs(api_srcs SRCS c10/core/Device.cpp)
 collect_srcs(api_srcs SRCS ATen/cuda/EmptyTensor.cpp)
 collect_srcs(api_srcs SRCS ATen/core/TensorMethods.cpp)
 collect_srcs(api_srcs SRCS ATen/AccumulateType.cpp)
+collect_srcs(api_srcs SRCS ATen/Utils.cpp)
 collect_srcs(api_srcs SRCS torch/csrc/api/include/torch/cuda.cpp)
 collect_srcs(api_srcs SRCS torch/library.cpp)
